+++ poem-openapi-derive/src/object.rs
@@ -63,6 +63,8 @@ struct ObjectArgs {
     #[darling(default)]
     deny_unknown_fields: bool,
     #[darling(default)]
+    example: bool,
+    #[darling(default)]
     external_docs: Option<ExternalDocument>,
     #[darling(default)]
     remote: Option<Path>,
@@ -313,6 +315,30 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
         None
     };
 
+    let (example, where_clause) = if args.example {
+        let new_where_clause = match where_clause {
+            Some(where_clause) => {
+                if where_clause.predicates.trailing_punct() {
+                    quote! { #where_clause Self: #crate_name::types::Example }
+                } else {
+                    quote! { #where_clause, Self: #crate_name::types::Example }
+                }
+            }
+            None => quote! { where Self: #crate_name::types::Example },
+        };
+        (
+            quote! {
+                <Self as #crate_name::types::ToJSON>::to_json(&#crate_name::types::Example::example())
+            },
+            new_where_clause,
+        )
+    } else {
+        (
+            quote! { ::std::option::Option::None },
+            quote!(#where_clause),
+        )
+    };
+
     let define_obj = quote! {
         impl #impl_generics #crate_name::types::Type for #ident #ty_generics #where_clause {
             const IS_REQUIRED: bool = true;
@@ -332,7 +358,9 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
             fn register(registry: &mut #crate_name::registry::Registry) {
                 registry.create_schema::<Self, _>(Self::name().into_owned(), |registry| {
                     #(#register_types)*
-                    #meta
+                    let mut meta = #meta;
+                    meta.example = #example;
+                    meta
                 })
             }